ARM: Orion: Hoist bridge interrupt handling out of the timer
The intent of this patch is to expose the other bridge cause
interrupts to users in the kernel.

- Add orion_bridge_irq_init to create a new edge triggered interrupt
  chip based on the bridge cause register
- Remove all interrupt register code from time.c and use normal
  interrupt functions instead
- Update the machines that use orion_time_init to call
  orion_bridge_irq_init and use the new signature

Tested on a Kirkwood platform.

I'm skeptical that MV78xx0 has a bridge interrupt cause/mask register,
it was setup so the timer code would touch those registers, so I've
preserved that, but prior to this patch the 'bridge cause register'
was only written to, never read. If it is wired-to-zero because it
doesn't exist then the timer will fail to function.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>