ARM: Orion wdt: Panic when the watchdog expires
If the platform device is setup with an interrupt handler then hook that
interrupt and call panic().

Currently the watchdog sets RSTOUTn, which triggers logic outside the
SOC to perform a reset, so most likely the panic will not complete.
However, if the external logic has a short delay then the panic will make
it out before the CPU reset.

Tested on a Kirkwood platform

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>