RDMA/uverbs: Ensure UVERBS_ATTR_TYPE_PTR_OUT zeros user memory
When the user passes a larger buffer, perhaps because it supports a
newer kernel, we want the kernel to write to the entire buffer,
even if it is just a zero fill at the end. This makes the API more
reliable as we've seen several cases where drivers in userspace forget
to zero buffers before passing to the kernel.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>