RDMA/uverbs: Copy the method and attribute specs into the UAPI header
This doesn't change what comes out of the UAPI header, but by copy the
text there it allows us to see all the type information and other data
from userspace.

This is the first step to remove the copies in the C file and have a single
source for the specs shared by all users. This will ultimately be done by
multi-including the uapi header with different macros sets in later patches.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>