RDMA/uverbs: Give prototypes to all uverbs data derived from the specs
This is the first step to using the specs language in the UAPI header to
directly create the supporting static const structures. To do this we need
regular names for all objects that cross compilation boundaries so that
the spec macros can link to them when they are expanded.

To avoid placing kernel internal symbol names in a uapi header all
linkages are created using macros based on the method, or object ID
string. The places in the source defining these objects must use the macro
to create the symbol name. Users of the macros can also no longer
be static as the reference will be from another compilation unit.

This also creates a different sort of clarity for the source. It is now
very clear what UAPI element each source object is linked to. Eg a
function implementation for a method will always note the method name in
the prototype.

The driver specific symbols for extensions to verbs will be handled
differently in later patches.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>