RDMA/uverbs: Provide a simple way for each object to remove itself
Lots of our objects are linked to underlying driver support and should
not appear in the parsing tree. Provide an object_supported hook for
the tree merger to call to determine if the object and all the enclosed
methods should be added to the tree.

FIXME: I didn't actually implement the merge bit of this

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>