RDMA/cm: Make cm_insert_remote_qpn() work on cm_id_private
Currently the cm_timewait_info is used to detect a condition when the
timewait_info is in the RB tree but not in the xarray.

However, this condition appears to be impossible.

The only places that assign to the cm_id_priv->timewait_info:

 cm_enter_timewait() - calls cm_remove_remote()
 cm_rest_to_idle() - calls cm_remove_remote()
 ib_send_cm_req() - State is IDLE so timewait_info == NULL (or this leaks
                    memory)
 cm_req_handler() - cm_id was just created by ib_create_cm_id()

And the only places that change the timewait_info->work.remote_id or
cm_id_priv->cm_id.remote_id:

 cm_req_handler() - calls cm_create_timewait_info() and uses the same id
                    for both
 cm_rep_handler() - calls cm_insert_remote_id() which requires the timewait
                    to not be in the RB tree already. Uses the same id for
                    both
 cm_sidr_req_handler() - sidr does not use timewait

Consider a present timewait but missing cm_id as a logical error, and
remove the handling for this from the code.

Have cm_insert_remote_qpn() setup the timewait data from the cm_id_priv,
add a priv->remote_ca_guid to make this possible.

Change-Id: I6c4d467c80782aa5e13ccb2da77d7eee905b1551
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>