RDMA/cm: Lock the cm_id after creation
ib_crate_cm_id() immediately places the id in the xarray, so it is visible
to network traffic.

The state is initially set to IB_CM_IDLE and all the MAD handlers will
test this state under lock and refuse to advance from IDLE, so adding
to the xarray is harmless.

However, all places that set id.state must do so under the spinlock to
ensure that any parallel network traffic sees a relase of all the stores
done here.

Thus we cannot omit the spinlock even though the object was just created.

Fixes: a977049dacde ("[PATCH] IB: Add the kernel CM implementation")
Change-Id: I63b10e479cbf833df8e62e28e3fe08eedaccf65f
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>