RDMA/core: Use an address_space to invalidate user VMAs in rdma_user_mmap
This is the same technique used by DRM, char/mem, CXL, etc

Each ib_uverbs_device creates an address space through an inode attached
to a filesystem and links that address space to every uverbs char device
FD.

The mm automatically keeps track of VMAs created by mmap on the uverbs
char device and links them to the inode's address space. A single call to
unmap_mapping_range() zaps the PTEs in every VMA, and the kernel ensures
this is zap is atomic with concurrent VMA modifications and dups. We
ensure new VMAs can not be created before unmap_mapping_range() is called.

Other than the lengthy boilerplate to create the address_space, this
dramatically simplifies this logic.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>