vfio/pci: Replace vfio_device_get_from_dev() with vfio_pci_get_drvdata()
We can usually directly get the vfio_pci_device associated with any pdev
by using the drvdata and the driver_lock(). The driver_lock prevents races
with remove, and the already present, check for bound PCI driver also
confirms the drvdata is of the right type.

Simply do these checks and return the result under the driver_lock(). The
refcount of the vfio_device cannot be zero at this point because the
attached driver proves that remove() hasn't run, and probe() has completed
fully.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>