vfio/fsl-mc: Use the driver_lock instead of reflck_lock
vfio_fsl_mc_reflck_attach() is trying to access the vdev affiliated with
a parent fsl_mc_device.

We can safely access the vdev from the drvdata so long as the driver_lock
is held. The driver_lock prevents racing with both probe() and remove()
call backs, which also hold the driver_lock, so we are certain to see a
fully setup and stable vdev.

Instead of searching, check that the correct driver is bound to the
parent, access the drvdata and then obtain the refcount on the reflck.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>