vfio: Move vfio_iommu_group_get() to vfio_register_group_dev()
We don't need to hold a reference to the group in the driver as well as
obtain a reference to the same group as the first thing
vfio_register_group_dev() does.

Since the drivers never use the group move this all into the core code.

FIXME: this is wrong error unwind afer vfio_iommu_group_get() doesn't
consider the noiommu case

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>