cover-letter: Finish adding type safety to vfio core and remove dead code
The last few APIs that accepted a 'struct device *' can now be eliminated
or switched to a 'struct vfio_device *'.

All of these were trying to get back to the group that the vfio_device
already has, so this transformation simplifies the logic and avoids a
bunch of work obtaining a group pointer that every driver already has.