vfio/platform: Remove vfio_platform_device->name
This is only used for diagnostic prints, instead consistently use the
sysfs name of the parent device.

All the prints were already showing the parent device sysfs name as the
'dev' is actually the parent device of the struct vfio_device.

Only AMBA gets a notable change where the old name was "vfio-amba-%08x" of
the parent and it will now be the sysfs name of the parent. The periphid
can be learned from the sysfs directory.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>