driver core: Pull required checks into driver_probe_device()
Checking if the dev is dead or if the dev is already bound is a required
precondition to invoking driver_probe_device(). All the call chains
leading here duplicate these checks.

Add it directly to driver_probe_device() so the precondition is clear and
remove the checks from device_driver_attach() and
__driver_attach_async_helper().

The other call chain going through __device_attach_driver() does have
these same checks but they are inlined into logic higher up the call stack
and can't be removed.

Preserve the sysfs uAPI behavior for store of succeeding if any driver is
already bound, but consistently fail if the device is unregistered or
dead.

Done in preparation for the next patches which will add additional
callers to driver_probe_device().

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>