RDMA/mlx5: Release transport domain if loopback enable failed
The failure to enable LB left transport domain allocated with wrong
number of LB properties (qps and user_td). That didn't reflect the
reality, because ucontext that was called to allocate this transport
domain flow returned failure to the user and wasn't created.

Refactor the code to have proper unwind.

Fixes: c85023e153e3 ("IB/mlx5: Add raw ethernet local loopback support")
Fixes: 0042f9e458a5 ("RDMA/mlx5: Enable vport loopback when user context or QP mandate")
Reported-by: Hans Petter Selasky <hanss@nvidia.com>
Reviewed-by: Huy Nguyen <huyn@nvidia.com>
Signed-off-by: Leon Romanovsky <leonro@nvidia.com>