vfio: Support migration drivers that do not implement P2P quiescence
When the VFIO_DEV_FLAGS_MIG_NOP2P flag is set the driver will not see
either of the P2P states. The FSM is flattened according to the
specification by treating the P2P's as their RUNNING varients. Requests to
move to a P2P state are rejected.

In this mode the driver need to implement 9 FSM arcs and the following
combination arcs are used:
                 PRE_COPY -> RUNNING -> STOP
                 PRE_COPY -> RUNNING -> STOP -> RESUMING
                 RESUMING -> STOP -> RUNNING
                 RESUMING -> STOP -> RUNNING -> PRE_COPY
                 RESUMING -> STOP -> STOP_COPY
                 RUNNING -> STOP -> RESUMING
                 RUNNING -> STOP -> STOP_COPY
                 STOP -> RUNNING -> PRE_COPY
                 STOP_COPY -> STOP -> RESUMING
                 STOP_COPY -> STOP -> RUNNING

This maps onto the specification by treating the P2P states as fully
running, so the new FSM arcs the driver implements are understood as:
                 PRE_COPY -> [PRE_COPY_P2P] -> STOP_COPY
                 RUNNING -> [RUNNING_P2P] -> STOP
                 STOP -> [RUNNING_P2P] -> RUNNING

Change-Id: Ib6d4551c5f9c7e406c18074295126ef64d8a13d8
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>