vfio: Have the core code decode the VFIO_DEVICE_FEATURE ioctl
Invoke a new device op 'device_feature' to handle just the data array
portion of the command. This lifts the ioctl validation to the core code
and makes it simpler for either the core code, or layered drivers, to
implement their own feature values.

Provide vfio_check_feature() to consolidate checking the flags/etc against
what the driver supports.

Change-Id: I8709758109d45f64ad7c0e17a0ae142bf4edfaed
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>