cover-letter: Make the rest of the VFIO driver interface use vfio_device
Prologue
========

This is series #4 in part of a larger work that arose from the minor
remark that the mdev_parent_ops indirection shim is useless and
complicates things.

Here we remove the use of 'struct device *' and 'struct vfio_group *' from
the rest of the driver facing VFIO interfaces.

I have been hoping Christoph's series would land by now to move kvmgt to
the new mdev interface, but that looks to be some more time. I've tried to
minimally clash with that work by adding a simple:

  struct vfio_device *mdev_legacy_get_vfio_device(struct mdev_device *mdev);

For kvmgt to use and get the vfio_device from its mdev_device. After
Christoph's series this API can be deleted too.
========

Prior series have transformed others parts of VFIO from working on struct
device or struct vfio_group into working directly on struct
vfio_device. Based on that work we now have vfio_device's readily
available in all the drivers.

Update the reset of the driver facing API to use vfio_device as an input.

The following are switched from struct device to struct vfio_device:
  vfio_register_notifier()
  vfio_unregister_notifier()
  vfio_pin_pages()
  vfio_unpin_pages()
  vfio_dma_rw()

The following group APIs are obsoleted and removed by just using struct
vfio_device with the above:
  vfio_group_pin_pages()
  vfio_group_unpin_pages()
  vfio_group_iommu_domain()
  vfio_group_get_external_user_from_dev()

To retain the performance of the new device APIs relative to their group
versions optimize how vfio_group_add_container_user() is used to avoid
calling it when the driver must already guarantee the device is open and
the container_users incrd.

The remaining exported VFIO group interfaces are only used by kvm.

This is also part of the iommufd work - moving the driver facing interface
to vfio_device provides a much cleaner path to integrate with iommufd.

issue: 2585478
Change-Id: I850be1196e6569658156e7718b87ee01859643c9
Cc: Christoph Hellwig <hch@lst.de>
Cc: intel-gfx@lists.freedesktop.org
Cc: intel-gvt-dev@lists.freedesktop.org
Cc: kvm@vger.kernel.org
Cc: "Tian, Kevin" <kevin.tian@intel.com>
Cc: "Liu, Yi L" <yi.l.liu@intel.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>